+++ jexxa-core/src/test/java/io/jexxa/application/domainservice/IJexxaPublisher.java
@@ -1,0 +1,9 @@
+package io.jexxa.application.domainservice;
+
+import io.jexxa.application.domain.valueobject.JexxaValueObject;
+
+public interface IJexxaPublisher
+{
+    void sendToQueue(JexxaValueObject jexxaValueObject);
+    void sendToTopic(JexxaValueObject jexxaValueObject);
+}